requirements.txt
asgiref==3.7.2
attrs==23.1.0
dj-database-url==2.1.0
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
dotenv==0.9.9
drf-spectacular==0.26.5
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
joblib==1.5.1
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
referencing==0.30.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.5.0
